requirements.txt
cookies==2.2.1
gevent==1.1b1
mock==1.2.0
pyramid==1.5.7
pyramid-jinja2==2.5
python-forecastio==1.3.4
